remove multi_dense
