--- extra_tests/regression_fits/multi_dense_316.json
@@ -1,95 +1,0 @@
-{
-  "preprocessing": [
-    {
-      "fl": "sng",
-      "smallx": 1.1168484687805176,
-      "largex": 1.4932247400283813,
-      "trainable": false
-    },
-    {
-      "fl": "g",
-      "smallx": 0.9923287034034729,
-      "largex": 3.7714877128601074,
-      "trainable": false
-    },
-    {
-      "fl": "v",
-      "smallx": 0.4898352324962616,
-      "largex": 1.672149658203125,
-      "trainable": false
-    },
-    {
-      "fl": "v3",
-      "smallx": 0.28436920046806335,
-      "largex": 3.4038245677948,
-      "trainable": false
-    },
-    {
-      "fl": "v8",
-      "smallx": 0.635401725769043,
-      "largex": 1.9356087446212769,
-      "trainable": false
-    },
-    {
-      "fl": "t3",
-      "smallx": 0.5063296556472778,
-      "largex": 2.4878852367401123,
-      "trainable": false
-    },
-    {
-      "fl": "t8",
-      "smallx": 0.8517391681671143,
-      "largex": 2.7349023818969727,
-      "trainable": false
-    },
-    {
-      "fl": "t15",
-      "smallx": 1.1255686283111572,
-      "largex": 1.68263578414917,
-      "trainable": false
-    }
-  ],
-  "stop_epoch": 1100,
-  "best_epoch": 1099,
-  "erf_tr": 2.4744620323181152,
-  "erf_vl": 2.905952215194702,
-  "chi2": 2.118803024291992,
-  "pos_state": "POS_VETO",
-  "arc_lengths": [
-    1.3125837362888904,
-    1.4889504508820537,
-    1.1794334718334645,
-    1.0006968075180218,
-    1.2461907760321072
-  ],
-  "integrability": [
-    0.0008942503627631937,
-    0.0008942503627628051,
-    5.014488272003792e-05,
-    0.46068788319826015,
-    0.007144232804419559
-  ],
-  "timing": {
-    "walltime": {
-      "Total": 11.391953229904175,
-      "start": 0.0,
-      "replica_set": 0.22628402709960938,
-      "replica_fitted": 11.391895294189453,
-      "replica_set_to_replica_fitted": 11.165611267089844
-    },
-    "cputime": {
-      "Total": 13.023389705,
-      "start": 0.0,
-      "replica_set": 0.2253089749999999,
-      "replica_fitted": 13.023330024,
-      "replica_set_to_replica_fitted": 12.798021048999999
-    }
-  },
-  "version": {
-    "keras": "3.6.0",
-    "tensorflow": "2.18.0, mkl=False",
-    "numpy": "1.26.4",
-    "nnpdf": "4.0.9.post1608.dev0+77c3343e3",
-    "validphys": "4.0.9.post1608.dev0+77c3343e3"
-  }
-}
