+++ n3fit/runcards/examples/Basic_runcard_parallel.yml
@@ -50,12 +50,8 @@ parameters: # This defines the parameter dictionary that is passed to the Model 
   threshold_chi2: 5.0
 
 fitting:
-  sum_rules: "All"
-  # NN23(QED) = sng=0,g=1,v=2,t3=3,ds=4,sp=5,sm=6,(pht=7)
-  # EVOL(QED) = sng=0,g=1,v=2,v3=3,v8=4,t3=5,t8=6,(pht=7)
-  # EVOLS(QED)= sng=0,g=1,v=2,v8=4,t3=4,t8=5,ds=6,(pht=7)
-  # FLVR(QED) = g=0, u=1, ubar=2, d=3, dbar=4, s=5, sbar=6, (pht=7)
   fitbasis: NN31IC # EVOL (7), EVOLQED (8), etc.
+  savepseudodata: False
   basis:
       # remeber to change the name of PDF accordingly with fitbasis
       # smallx, largex: preprocessing ranges
