+++ n3fit/src/n3fit/backends/__init__.py
@@ -19,6 +19,6 @@ from n3fit.backends.keras_backend.internal_state import (
     set_eager,
     set_initial_state,
 )
-from n3fit.backends.keras_backend.multi_dense import MultiInitializer
+from n3fit.backends.keras_backend.multi_initializer import MultiInitializer
 
 print("Using Keras backend")
